RALPH: Delete workspace-mock-data — closure-isolate the four splits (#250)
Final cleanup of the workspace split. With profile (#246),
workspace-employees (#247), invitations (#248), and company-info (#249)
each living behind the data-source seam, the legacy `workspace-mock-data`
file no longer carries production state — every consumer reaches it only
through one of the four in-memory adapters. This slice dissolves the
shared singleton: each adapter gets its own closure-isolated store.

Adapter changes
- src/data/clients/profile-in-memory.ts: rewritten as a
  closure-isolated factory. State (`me`, `settings`) lives in the
  closure; `me` / `settings` / `update` / `changePassword` operate on
  it directly. Defaults `DEFAULT_ME` / `DEFAULT_SETTINGS` inlined from
  the legacy `SEED_ME` / `SEED_USER_SETTINGS`.
- src/data/clients/company-info-in-memory.ts: rewritten as
  closure-isolated. `get()` reads the closure-held record. `info`
  option seeds it; defaults to `{ name: "ОРМАТЕК" }`.
- src/data/clients/workspace-employees-in-memory.ts: rewritten as
  closure-isolated. The roster + id counter for invitees live in the
  closure. Throws `NotFoundError` directly from the new `requireIndex`
  helper instead of routing through the old `translateNotFound` shim
  (the legacy mock threw a generic Error, the new one throws the
  typed error itself). Cross-store reach-in to
  `_getCompanySummariesByIds` from companies-mock-data preserved at
  invite time so an invitee's `companies` array stays coherent with
  the companies adapter — the only cross-store call left in this
  domain. The seed roster moves to `src/data/seeds/workspace-employees.ts`
  (mirrors seeds/companies + seeds/items + seeds/emails).
- src/data/clients/invitations-in-memory.ts: doc comment trimmed —
  no longer references the still-pending workspace-mock-data
  deletion now that it's done.

Domain types
- src/data/domains/profile.ts: types defined inline instead of
  re-exporting from workspace-mock-data. `UserSettings`,
  `SettingsPatch`, `ChangePasswordResponse`, `CurrentEmployee`. Imports
  `EmployeeRole` from types.ts for the `CurrentEmployee.role` field.
- src/data/domains/workspace-employees.ts: types defined inline
  instead of re-exporting. `WorkspaceEmployee`, `WorkspaceEmployeeDetail`,
  `InviteEmployeeData`, `UpdateWorkspaceEmployeeData`,
  `UpdatePermissionsData`. Imports underlying primitives
  (`CompanySummary`, `Employee`, `EmployeePermissions`, `EmployeeRole`,
  `PermissionLevel`) from types.ts.

Seeds
- src/data/seeds/workspace-employees.ts (new): SEED_WORKSPACE_EMPLOYEES
  — the 7-employee roster (admin + 6 others, spanning roles and
  registration states). MAIN_COMPANY_SUMMARY moved here too. IDs
  1–4 still overlap with the main tenant's company-1 employees in
  SEED_COMPANIES so cross-domain data stays coherent.

Contract test updates
- src/data/clients/profile-contract.test.ts: dropped
  `_resetWorkspaceStore` / `_setMe` / `_setUserSettings` reach-ins.
  Memory adapter now seeds via the constructor option
  `createInMemoryProfileClient({ me, settings })`.
- src/data/clients/company-info-contract.test.ts: dropped
  `_resetWorkspaceStore` / `_setCompanyInfo`. Memory adapter seeds
  via `createInMemoryCompanyInfoClient({ info })`.
- src/data/clients/workspace-employees-contract.test.ts: dropped
  `_resetWorkspaceStore` / `_setWorkspaceEmployees`. Memory adapter
  seeds via `createInMemoryWorkspaceEmployeesClient({ seed })`.

Deletions
- src/data/workspace-mock-data.ts (409 lines) deleted.
- src/data/workspace-mock-data.test.ts (221 lines) deleted —
  its assertions are subsumed by the per-domain Layer A hook tests
  and Layer B contract tests already in place.

CONTEXT.md
- In-memory adapter blurb updated: profile, company-info,
  workspace-employees, and invitations are now in the
  closure-isolated list (alongside companies + emails). Items,
  suppliers, tasks, folders, and notifications remain
  singleton-wrapping until cross-entity rules move to the
  procurement-operations module (#251). Workspace-employees'
  `_getCompanySummariesByIds` reach-in to companies-mock-data is
  called out explicitly so future readers see why it's the only
  cross-store call left in that domain.

Verified
- pnpm typecheck / lint / test / knip / build all clean.
- 1426 tests passing (was 1445; net −19 from deleting the legacy
  workspace-mock-data.test.ts whose assertions are subsumed by the
  per-domain hook + contract tests).

Acceptance criteria per #250
- No file under src/ imports from workspace-mock-data ✓
  (verified via grep — only stale doc comments remained, those
  trimmed too)
- workspace-mock-data.ts and workspace-mock-data.test.ts deleted ✓
- pnpm typecheck, pnpm lint, pnpm knip all pass ✓
- Existing tests still pass ✓

Notes for next iteration
- Workspace-employees still has one cross-store reach-in
  (`_getCompanySummariesByIds`). Lifting it would mean injecting a
  `CompaniesClient` (or a slim `getSummaries` port) into the
  workspace-employees factory — a clean follow-up but not strictly
  required by #250's AC.
- Profile / company-info / workspace-employees / invitations now
  match the precedent set by companies + emails: closure-isolated,
  no shared store, no underscore reach-ins from consumers.
- Items / suppliers / tasks / folders / notifications remain
  singleton-wrapping; they dissolve as part of #251
  (procurement-operations) — by the time cross-entity rules move
  to the named seam, no callers need the singleton's underscore
  helpers.

Closes #250.
