+++ src/data/clients/company-info-in-memory.ts
@@ -1,27 +1,26 @@
 import type { CompanyInfo } from "../domains/company-info";
-import { _setCompanyInfo, fetchCompanyInfoMock } from "../workspace-mock-data";
+import { delay } from "../mock-utils";
 import type { CompanyInfoClient } from "./company-info-client";
 
+const DEFAULT_INFO: CompanyInfo = { name: "ОРМАТЕК" };
+
 export interface InMemoryCompanyInfoOptions {
-	/** Replace the module-level mock store at construction time. Tests pass
-	 * this to land on a known workspace identity (e.g. "Acme Corp") without
-	 * reaching into `_setCompanyInfo` directly. */
+	/** Replace the seeded workspace identity (e.g. "Acme Corp"). */
 	info?: CompanyInfo;
 }
 
 /**
- * Build an in-memory company-info adapter wrapping the module-level workspace
- * mock store. Singleton-wrapping (rather than closure isolation) is the right
- * shape here because `workspace-mock-data` is shared with the profile and
- * workspace-employees domains until #250 dissolves it. Once those splits are
- * cleaned up, this can become closure-isolated.
+ * Build a closure-isolated in-memory company-info adapter. State (the active
+ * workspace's identity record) lives in the closure — every call to the
+ * factory produces an independent store.
  */
 export function createInMemoryCompanyInfoClient(options?: InMemoryCompanyInfoOptions): CompanyInfoClient {
-	if (options?.info !== undefined) _setCompanyInfo(options.info);
+	const info: CompanyInfo = { ...(options?.info ?? DEFAULT_INFO) };
 
 	return {
 		async get(): Promise<CompanyInfo> {
-			return fetchCompanyInfoMock();
+			await delay();
+			return { ...info };
 		},
 	};
 }
