+++ src/data/domains/workspace-employees.ts
@@ -1,3 +1,5 @@
+import type { CompanySummary, Employee, EmployeePermissions, EmployeeRole, PermissionLevel } from "../types";
+
 /**
  * Workspace-employees domain types — the workspace's full employee roster.
  * Backs `useWorkspaceEmployees`, `useWorkspaceEmployeeDetail`,
@@ -10,15 +12,37 @@
  * and owns invitation lifecycle. The two domains share the underlying
  * `Employee` shape but model different things — a workspace employee carries a
  * `companies: CompanySummary[]` array because they may belong to several.
- *
- * The types are re-exported from the legacy `workspace-mock-data` until that
- * file is dissolved in #250; consumers should import them from this module so
- * the eventual deletion is a single import-rename pass.
  */
-export type {
-	InviteEmployeeData,
-	UpdatePermissionsData,
-	UpdateWorkspaceEmployeeData,
-	WorkspaceEmployee,
-	WorkspaceEmployeeDetail,
-} from "../workspace-mock-data";
+
+export interface WorkspaceEmployee extends Employee {
+	companies: CompanySummary[];
+}
+
+export type WorkspaceEmployeeDetail = WorkspaceEmployee & { permissions: EmployeePermissions };
+
+export interface InviteEmployeeData {
+	email: string;
+	firstName: string;
+	lastName: string;
+	patronymic: string;
+	position: string;
+	role: EmployeeRole;
+	companies: string[];
+}
+
+export interface UpdateWorkspaceEmployeeData {
+	firstName?: string;
+	lastName?: string;
+	patronymic?: string;
+	position?: string;
+	role?: EmployeeRole;
+	phone?: string;
+}
+
+export interface UpdatePermissionsData {
+	procurement?: PermissionLevel;
+	tasks?: PermissionLevel;
+	companies?: PermissionLevel;
+	employees?: PermissionLevel;
+	emails?: PermissionLevel;
+}
