+++ src/data/seeds/workspace-employees.ts
@@ -1,0 +1,173 @@
+import type { WorkspaceEmployeeDetail } from "../domains/workspace-employees";
+import type { CompanySummary } from "../types";
+
+const MAIN_COMPANY_SUMMARY: CompanySummary = {
+	id: "company-1",
+	name: "ОРМАТЕК",
+	isMain: true,
+	addresses: [
+		{
+			id: "addr-c1-office",
+			name: "Головной офис",
+			address: "г. Москва, Ленинградское шоссе, д. 16А, стр. 1",
+			isMain: true,
+		},
+	],
+	employeeCount: 4,
+	procurementItemCount: 8,
+};
+
+/**
+ * Default workspace-employees roster — admin + 6 others, spanning roles and
+ * registration states. IDs 1–4 overlap with the main tenant's company-1
+ * employees (`SEED_COMPANIES`) so the data stays coherent across the seam.
+ */
+export const SEED_WORKSPACE_EMPLOYEES: WorkspaceEmployeeDetail[] = [
+	{
+		id: 1,
+		firstName: "Иван",
+		lastName: "Журавлёв",
+		patronymic: "Сергеевич",
+		position: "Директор по закупкам",
+		role: "admin",
+		phone: "+79161000001",
+		email: "ivan.zhuravlyov.58@mostholding.ru",
+		registeredAt: "2024-01-15T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-1",
+			employeeId: 1,
+			procurement: "edit",
+			tasks: "edit",
+			companies: "edit",
+			employees: "edit",
+			emails: "edit",
+		},
+	},
+	{
+		id: 2,
+		firstName: "Ольга",
+		lastName: "Соколова",
+		patronymic: "Андреевна",
+		position: "Руководитель отдела закупок",
+		role: "admin",
+		phone: "+79161000002",
+		email: "o.sokolova@ormatek.com",
+		registeredAt: "2024-02-01T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-2",
+			employeeId: 2,
+			procurement: "edit",
+			tasks: "edit",
+			companies: "edit",
+			employees: "edit",
+			emails: "edit",
+		},
+	},
+	{
+		id: 3,
+		firstName: "Дмитрий",
+		lastName: "Орлов",
+		patronymic: "Михайлович",
+		position: "Менеджер по закупкам",
+		role: "user",
+		phone: "+79161000003",
+		email: "d.orlov@ormatek.com",
+		registeredAt: "2024-03-12T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-3",
+			employeeId: 3,
+			procurement: "edit",
+			tasks: "edit",
+			companies: "view",
+			employees: "view",
+			emails: "view",
+		},
+	},
+	{
+		id: 4,
+		firstName: "Екатерина",
+		lastName: "Белова",
+		patronymic: "Игоревна",
+		position: "Аналитик отдела закупок",
+		role: "user",
+		phone: "+79161000004",
+		email: "e.belova@ormatek.com",
+		registeredAt: "2024-05-20T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-4",
+			employeeId: 4,
+			procurement: "view",
+			tasks: "view",
+			companies: "view",
+			employees: "none",
+			emails: "none",
+		},
+	},
+	{
+		id: 5,
+		firstName: "Николай",
+		lastName: "Фёдоров",
+		patronymic: "Олегович",
+		position: "Категорийный менеджер",
+		role: "user",
+		phone: "+79161000005",
+		email: "n.fedorov@ormatek.com",
+		registeredAt: "2024-08-03T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-5",
+			employeeId: 5,
+			procurement: "edit",
+			tasks: "edit",
+			companies: "view",
+			employees: "none",
+			emails: "view",
+		},
+	},
+	{
+		id: 6,
+		firstName: "Анна",
+		lastName: "Смирнова",
+		patronymic: "Николаевна",
+		position: "Финансовый аналитик",
+		role: "user",
+		phone: "+79161000006",
+		email: "a.smirnova@ormatek.com",
+		registeredAt: "2025-01-10T10:00:00Z",
+		companies: [MAIN_COMPANY_SUMMARY],
+		permissions: {
+			id: "perm-w-6",
+			employeeId: 6,
+			procurement: "view",
+			tasks: "view",
+			companies: "none",
+			employees: "none",
+			emails: "none",
+		},
+	},
+	{
+		id: 7,
+		firstName: "Дмитрий",
+		lastName: "Попов",
+		patronymic: "",
+		position: "Менеджер по закупкам",
+		role: "user",
+		phone: "",
+		email: "d.popov@ormatek.com",
+		registeredAt: null,
+		companies: [],
+		permissions: {
+			id: "perm-w-7",
+			employeeId: 7,
+			procurement: "none",
+			tasks: "none",
+			companies: "none",
+			employees: "none",
+			emails: "none",
+		},
+	},
+];
